# deliverapi/templates/lib/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Auto Deployment System API",
        "description": "Move your app forward with the Auto Deployment System API",
        "version": "v1.0"
    },
    "host": "127.0.0.1:8000",
    "schemes": [
        "http"
    ],
    "basePath": "",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/deployment/": {
            "get": {
                "summary": "List all Deployments history",
                "description": " Return a list of deployment history",
                "tags": [
                    "Deployment"
                ],
                "responses": {
                    "200": {
                        "description": "A list of deployment history"
                    },
                    "403": {
                        "description": "Authentication credentials were not provided"
                    }
                }
            },
          "post": {
            "summary": " create a new deployment instance",
            "description": "Submit application version and code version , create a new deployment instance.",
            "tags": [
                "Deployment"
            ],
            "responses": {
                "201": {
                    "description": "successful operation"
                },
                "400": {
                    "description": "Invalid parameters supplied"
                },
                "403": {
                    "description": "Authentication credentials were not provided"
                }
            },
            "parameters":[
              {
                "name": "codeversion",
                "in": "formData",
                "required": "true",
                "description": "version from the remote version library ",
                "type": "string"
              },
              {
                "name": "appversion",
                "in": "formData",
                "required": "true",
                "description": "application version",
                "type": "string"
              }
            ]

          }
        },
...
    },
...
}